Update index.php
Changed the url function to determine if the front-end server is using HTTPS.  When using a reverse proxy server that does SSL offloading the webserver is unaware that HTTPS is served and produces HTTP links resulting in mixed content.  This change will automatically detect if there is a final HTTPS protocol and form urls accordingly.